OEM - anushthaFatehPortal/backend/requirements.txt
# UnityESS TCP — FastAPI Backend Dependencies
# ===================================================

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.12

# Database
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
alembic==1.13.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-keycloak==4.2.2

# Task queue
celery==5.4.0
redis==5.2.0

# Object storage
boto3==1.35.0          # MinIO via S3 API

# AI extraction
anthropic==0.40.0
PyPDF2==3.0.1          # PDF text extraction
google-generativeai==0.8.3  # Gemini API (free)

# Document generation
weasyprint==63.0       # PDF
openpyxl==3.1.5        # Excel
python-pptx==1.0.2     # PPTX

# Validation & config
pydantic==2.9.2
pydantic-settings==2.6.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
tenacity==9.0.0

# Dev / testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
